requirements.txt
astroid==2.3.2
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.0
dnspython==1.16.0
eventlet==0.25.1
eyeD3==0.8.10
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mutagen==1.42.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
pyOpenSSL==19.1.0
python-engineio==3.11.1
python-magic==0.4.15
python-socketio==4.4.0
python-vlc==3.0.7110
requests==2.22.0
six==1.12.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
